requirements.txt
eth-account
hyperliquid-python-sdk
